Reorganized CSS files
