+++ frontend/src/styling/components/CharacterQuickItemsComp.css
