+++ frontend/src/styling/components/CharacterStatsComp.css
