+++ frontend/src/styling/pages/CharacterListPage.css
