+++ frontend/src/styling/pages/CharacterPage.css
