+++ frontend/src/styling/pages/page-tabs/CharacterInventoryTab.css
