+++ frontend/src/styling/pages/page-tabs/CharacterSpellcastingTab.css
